Reorganize the compress-texture package
