+++ bundles/pixi.js/package.json
@@ -40,6 +40,7 @@
   "dependencies": {
     "@pixi/accessibility": "5.3.2",
     "@pixi/app": "5.3.2",
+    "@pixi/compressed-textures": "5.3.2",
     "@pixi/constants": "5.3.2",
     "@pixi/core": "5.3.2",
     "@pixi/display": "5.3.2",
@@ -53,7 +54,6 @@
     "@pixi/graphics": "5.3.2",
     "@pixi/interaction": "5.3.2",
     "@pixi/loaders": "5.3.2",
-    "@pixi/compressed-textures": "5.3.2",
     "@pixi/math": "5.3.2",
     "@pixi/mesh": "5.3.2",
     "@pixi/mesh-extras": "5.3.2",
