+++ packages/compressed-textures/src/index.ts
@@ -1,9 +1,3 @@
-import { CompressedTextureLoader } from './CompressedTextureLoader';
-import { DDSLoader } from './DDSLoader';
-import { KTXLoader } from './KTXLoader';
-
-export {
-    CompressedTextureLoader,
-    DDSLoader,
-    KTXLoader
-};
+export * from './const';
+export * from './resources';
+export * from './loaders';
