+++ packages/compressed-textures/src/loaders/CompressedTextureLoader.ts
@@ -1,7 +1,7 @@
 import { LoaderResource } from '@pixi/loaders';
 
 import type { Loader } from '@pixi/loaders';
-import type { INTERNAL_FORMATS } from '@pixi/constants';
+import type { INTERNAL_FORMATS } from '../const';
 import type { ILoaderResource } from '@pixi/loaders';
 
 /**
