+++ packages/compressed-textures/src/loaders/DDSLoader.ts
@@ -1,7 +1,7 @@
-import { CompressedTextureResource } from '@pixi/core';
-import { INTERNAL_FORMATS, INTERNAL_FORMAT_TO_BYTES_PER_PIXEL } from '@pixi/constants';
+import { CompressedTextureResource } from '../resources';
+import { INTERNAL_FORMATS, INTERNAL_FORMAT_TO_BYTES_PER_PIXEL } from '../const';
 import { LoaderResource } from '@pixi/loaders';
-import { registerCompressedTextures } from './registerTextures';
+import { registerCompressedTextures } from './registerCompressedTextures';
 
 import type { ILoaderResource } from '@pixi/loaders';
