+++ packages/compressed-textures/src/loaders/index.ts
@@ -1,0 +1,3 @@
+export * from './CompressedTextureLoader';
+export * from './DDSLoader';
+export * from './KTXLoader';
