+++ packages/compressed-textures/src/loaders/registerCompressedTextures.ts
@@ -1,9 +1,7 @@
 import { MIPMAP_MODES, ALPHA_MODES } from '@pixi/constants';
 import { BaseTexture, Texture } from '@pixi/core';
 
-/* eslint-disable @typescript-eslint/no-unused-vars */
-import type { CompressedTextureResource } from '@pixi/core';
-/* eslint-enable @typescript-eslint/no-unused-vars */
+import type { CompressedTextureResource } from '../resources/CompressedTextureResource';
 
 /**
  * Creates base-textures and textures for each compressed-texture resource and adds them into the global
