+++ packages/compressed-textures/src/resources/index.ts
@@ -1,0 +1,2 @@
+export * from './BlobResource';
+export * from './CompressedTextureResource';
