+++ packages/core/src/textures/resources/index.ts
@@ -25,10 +25,8 @@ INSTALLED.push(
 export * from './autoDetectResource';
 export * from './AbstractMultiResource';
 export * from './ArrayResource';
-export * from './BlobResource';
 export * from './BufferResource';
 export * from './CanvasResource';
-export * from './CompressedTextureResource';
 export * from './CubeResource';
 export * from './ImageResource';
 export * from './SVGResource';
